# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Maintenance Task & QC Management System API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: assignedTo
          in: query
          required: false
          schema:
            type: integer
        - name: sectionId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID with full detail
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail with time entries and QC reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      summary: Update task status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStatusInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/time:
    get:
      operationId: listTimeEntries
      tags: [time]
      summary: List time entries for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: startTimeTracking
      tags: [time]
      summary: Start time tracking (creates an open entry)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartTimeInput"
      responses:
        "201":
          description: Created time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /tasks/{taskId}/time/pause:
    post:
      operationId: pauseTimeTracking
      tags: [time]
      summary: Pause active time tracking (requires reason)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseTimeInput"
      responses:
        "200":
          description: Paused time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /tasks/{taskId}/time/resume:
    post:
      operationId: resumeTimeTracking
      tags: [time]
      summary: Resume time tracking (creates new open entry)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: New time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        assetId:
          type: integer
        assetName:
          type: string
        sectionId:
          type: integer
        sectionName:
          type: string
        stageId:
          type: integer
        stageName:
          type: string
        stageNumber:
          type: integer
        bladeCountMin:
          type: integer
        bladeCountMax:
          type: integer
        componentId:
          type: integer
        componentName:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        createdById:
          type: integer
        estimatedHours:
          type: string
        deadline:
          type: string
          format: date-time
        priority:
          type: string
          enum: [high, medium, low]
        status:
          type: string
          enum: [draft, assigned, in_progress, paused, submitted, under_qc, approved, rejected, overdue]
        totalMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, createdAt]
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            timeEntries:
              type: array
              items:
                $ref: "#/components/schemas/TimeEntry"
            qcReviews:
              type: array
              items:
                $ref: "#/components/schemas/QcReview"
            activeTimeEntry:
              $ref: "#/components/schemas/TimeEntry"
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        pauseReason:
          type: string
        isActive:
          type: boolean
      required: [id, taskId, userId, startTime, isActive]
    QcReview:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, taskId, reviewerId, decision, createdAt]
...
    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assetId:
          type: integer
        sectionId:
          type: integer
        stageId:
          type: integer
        componentId:
          type: integer
        assignedToId:
          type: integer
        estimatedHours:
          type: number
        deadline:
          type: string
          format: date-time
        priority:
          type: string
          enum: [high, medium, low]
      required: [title, assetId, sectionId, priority]
    UpdateTaskStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [draft, assigned, in_progress, paused, submitted, under_qc, approved, rejected, overdue]
        pauseReason:
          type: string
        qcComment:
          type: string
      required: [status]
    StartTimeInput:
      type: object
      properties: {}
    PauseTimeInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...